# docs/api設計書/openapi.yml
openapi: 3.0.3
info:
  title: Hono Template API
  description: A REST API for managing todos and users with authentication
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /users:
    get:
      tags:
        - users
      summary: Get all users
      description: Retrieve a list of all users (admin only)
      operationId: getUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  meta:
                    $ref: "#/components/schemas/ListMeta"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessError"
        "403":
          description: Forbidden - admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessError"
              example:
                error:
                  code: "FORBIDDEN"
                  message: "Admin access required"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessError"
  /users/{id}:
    get:
      tags:
        - users
      summary: Get user by ID
      description: Retrieve a specific user by ID
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessError"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessError"
              example:
                error:
                  code: "USER_NOT_FOUND"
                  message: "No user found with the provided ID"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessError"

    put:
      tags:
        - users
      summary: Update user
      description: Update user information
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Jane Doe"
                email:
                  type: string
                  format: email
                  example: "jane@example.com"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessError"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessError"

    delete:
      tags:
        - users
      summary: Delete user
      description: Delete a user account
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  deletedId:
                    type: string
                    example: "123e4567-e89b-12d3-a456-426614174000"
                  success:
                    type: boolean
                    example: true
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessError"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "John Doe"
        role:
          type: string
          enum: [user, admin]
          example: "user"
        createdAt:
          type: string
          format: date-time
          example: "2025-07-01T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-07-01T10:00:00Z"
...
    ListMeta:
      type: object
      properties:
        totalCount:
          type: integer
          description: Total number of items
          example: 50
        page:
          type: integer
          description: Current page number
          example: 1
        limit:
          type: integer
          description: Number of items per page
          example: 10
        totalPages:
          type: integer
          description: Total number of pages
          example: 5
        hasNext:
          type: boolean
          description: Whether there is a next page
          example: true
        hasPrev:
          type: boolean
          description: Whether there is a previous page
          example: false
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                description: Error code
                example: "VALIDATION_ERROR"
              message:
                type: string
                description: Error message
                example: "Email field is required"
              field:
                type: string
                description: Field name
                example: "email"
    BusinessError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
              example: "UNAUTHORIZED"
            message:
              type: string
              description: Error message
              example: "Invalid credentials"